Add permissions to page object
Now, much like an APIResponse, we can now evaluate permissions for mates
based on an array of function calls.  This also adds enforcement of
permissions to /edit/ and /create/ namespaces.

Also restructures the home page in preparation for more content there.

TestTemplate was changed as a proof-of-concept to be a endpoint generator.
That is, it consumes the URI class' array of tags and generates route responders
for each, by default redirecting to /404/.