Create gallery layout, better endpoints for gals
Now there are API endpoints to add images to galleries. As well as
front-end js to support adding images with the media manager.

TODO: enforce priviliges!! The mediaManager doesn't need to be
added to the page if the user isn't logged in, and the JS
that facilitates the edit button doesn't need to be there either.
Or it needs to be disabled.