Make a public gallery for selfies
It's actually an entirely new endpoint that uploades AND
adds it to the sacrificial user's gallery (the selfies user).
When deploying this on a live machine, remember to update the site-settings
to hold the gallery id of the sacrificial gallery, and the userid of the
sacrifical user.

This also updates the error handling for all uses of the API so that all
errors are shown instead of just the first one.