Basics for media sizes
Create table schema, create queueworker script,
add functions to the MediaManager to enqueue and then to process
jobs, add dependencies to composer for image processing,
create exec scripts for testing things.

Currently the MediaManager function to process/generate sizes
just opens the file and doesn't do anything with it. This is just
a proof-of concept so that we can see that it's working.