Make it work!!
Okay, so now, sizes can actually be generated and saved.
Calling `php Exec/FakeImageWorker.php 14` will (right then) generate
the medium and small sizes for that image. It will not, however, create
a record in the mysql table (not yet).

Exec/FilesystemTest.php lists the contents of the filesystem that's registered
in MediaManager.

Libs/MediaProcessor.php is now a static class of functions for performing image
generation.

TODO:
   Create records of created images in MySql
   Automatically enqueue jobs when images are uploaded.