Add support for other sizes to queries
This includes returning the srcs of other sizes as part of
API responses, and using that size in the MediaManager as well.

(In the MediaManager, when you upload, the other sizes have not been
generated yet, so it will fall back to the original size if the small
is not available.)

TODO: Enhance the Content object so that it has better getter functions
for the leader image. Currently we use $post->leader directly. Perhaps
add a function that's $post->("small"); that will try to get the small
image but will still return *something* if the small is not available,
IE, the next-size-up.